pyproject.toml
[project]
name = "try01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "pandas",
  "geopandas",
  "matplotlib",
  "seaborn",
  "folium",
  "earthengine-api",
  "ipykernel",
  "geemap",
  "scipy",
  "pytest",
  "pyyaml",
  "fastapi",
  "uvicorn",
  "httpx",
]

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
addopts = "-q"
